# docs/swagger.yaml
definitions:
  albums.AddImagesToAlbumRequest:
    properties:
      image_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - image_ids
    type: object
  albums.AlbumDTO:
    properties:
      cover_url:
        type: string
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
      image_count:
        type: integer
      name:
        type: string
      updated_at:
        type: integer
    type: object
  albums.AlbumDetailResponse:
    properties:
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
      image_count:
        type: integer
      images:
        items:
          $ref: '#/definitions/albums.AlbumImageDTO'
        type: array
      name:
        type: string
      updated_at:
        type: integer
    type: object
  albums.AlbumImageDTO:
    properties:
      created_at:
        type: integer
      file_size:
        type: integer
      height:
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      original_name:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
  albums.CreateAlbumResponse:
    properties:
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: integer
    type: object
  albums.ListAlbumsResponse:
    properties:
      albums:
        items:
          $ref: '#/definitions/albums.AlbumDTO'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  albums.UpdateAlbumRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  albums.UpdateAlbumResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: integer
    type: object
  albums.createAlbumRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/albums:
    get:
      consumes:
      - application/json
      description: Get paginated list of user's albums with image counts and cover
        URLs
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Album list
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/albums.ListAlbumsResponse'
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: List albums
      tags:
      - albums
    post:
      consumes:
      - application/json
      description: Create a new album with name and optional description
      parameters:
      - description: Album creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/albums.createAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Album created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/albums.CreateAlbumResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Create album
      tags:
      - albums
  /api/v1/albums/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an album by ID (images in the album will not be deleted)
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Album deleted successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid album ID format
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete album
      tags:
      - albums
    get:
      consumes:
      - application/json
      description: Get detailed information about an album including all images
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Album details
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/albums.AlbumDetailResponse'
              type: object
        "400":
          description: Invalid album ID format
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Get album detail
      tags:
      - albums
    put:
      consumes:
      - application/json
      description: Update album name and description
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Album update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/albums.UpdateAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Album updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/albums.UpdateAlbumResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Update album
      tags:
      - albums
  /api/v1/albums/{id}/images:
    post:
      consumes:
      - application/json
      description: Add multiple images to an existing album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image IDs to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/albums.AddImagesToAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Images added to album successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album or images not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Add images to album
      tags:
      - albums
  /api/v1/albums/{id}/images/{imageId}:
    delete:
      consumes:
      - application/json
      description: Remove a single image from an album (image itself is not deleted)
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image ID
        in: path
        name: imageId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Image removed from album successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album or image not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove image from album
      tags:
      - albums
...
swagger: "2.0"